Arguments

Arguments

Arguments

data
A dataframe or tibble.

sql_statement
A SQL Statement.

sql_statements
A list or vector of SQL Statements of length 1 or greater.

schema
The target schema for the operation.

tableName
The target table in the schema for an operation or for writing functions, the name of the table that will be created.

verbose
If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.

render_sql
If TRUE, the SQL statement for the operation is returned in the console.

predicate
Evaluating function.

stop_message
Message to receive with the predicate evaluates to true.

arguments
(option) character vector of arguments for the new function. If missing, the new function will not have any arguments. Arguments in the predicate should also be considered. For example, if the predicate is "is.logical(x)", the arguments value should be "x".

.data
A dataframe or tibble.

db
name of database

sqlQuery
sql statement of length 1 that will be the sole key for cache retrieval.

object
R object to cache.

dirs
Argument passed to saveCache, findCache, and/or loadCache

remove
If TRUE, the Connection object argument is removed from the parent environment.

vector
vector of values that the SQL query is filtering for

field
Single field to be filtered for

term
Character string that the field parameter is searched for.

warn_message
Message to receive with the predicate evaluates to true.

dbname
Name of a local Postgres database, Default: 'athena'

port
The port on the server to connect to, Default: 5432

datetime_field
PARAM_DESCRIPTION

log_table
PARAM_DESCRIPTION

conn
Connection object

file
path to sql file

warn_no_rows
If TRUE, a warning is given that query has returned 0 rows.

cascade
If TRUE, a DROP SCHEMA CASCADE is performed.

group
group name, Default: NULL

user
user name, Default: NULL

fields
Fields selected for. Defaults to "*".

distinct
If TRUE, the distinct row count will be returned.

resultTableName
Table that the final output will be written to.

sqlList
list object of queries

progressBar
If TRUE, a progress bar is returned in the console.

skipErrors
If TRUE, if a SQL execution returns an error, the statement is printed back in red in the console and the iteration will proceed to the next sql statement in line.

errorFile
(optional) path to the error file where any errors are written to if not null.

if_exists
If TRUE, the table will be dropped only if it exists.

case_insensitive
If TRUE, both sides of the query are converted to lowercase.

values
Vector of length 1 or greater to search for.

instSubdir
Name of subdirectory in the inst/ folder

FileName
Name of file in subdirectory

package
Package name

table
Target table for the operation.

select_table_fields
The fields to select for in the first table

select_join_on_fields
The fields to select for in the table being join to the first table

column
Column to join on.

join_on_schema
Schema of the table that is being joined to the first table.

join_on_table
Table that is being joined to the first table.

join_on_column
Column in the join_on_table that is joined onto the column of the first table.

kind
Type of join. Defaults to left, and options include "LEFT", "RIGHT", "INNER", or "FULL"

where_in_field
Paired with where_in_vector, adds a "WHERE field IN (vector)" clause to the query corresponding to the field-value pair.

where_not_in_field
Paired with where_not_in_vector, adds a "WHERE field NOT IN (vector)" clause to the query corresponding to the field-value pair.

limit
(Optional) Integer of the row limit. Takes precedence over random if both arguments are provided.

random
(Optional) Integer of the random number of rows to return. Is preceded by limit if both limit and random are provided.