Connect to a Local Postgres Database

Connect and disconnect to a local Postgres database that does not require credentialing. These functions automatically set the dbms and server arguments to "postgresql" and "localhost/dbname", leaving the database name (dbname) and port (port) as the only connection details that are modifiable by the user. Note that these functions cannot be used to connect to a local instance that requires a username and password.

Connect and disconnect to a local Postgres database that does not require credentialing. These functions automatically set the dbms and server arguments to "postgresql" and "localhost/dbname", leaving the database name (dbname) and port (port) as the only connection details that are modifiable by the user. Note that these functions cannot be used to connect to a local instance that requires a username and password.

local_connect(
  dbname = "athena",
  port = 5432,
  user = NULL,
  password = NULL,
  extraSettings = NULL,
  oracleDriver = "thin",
  connectionString = NULL,
  pathToDriver = system.file(package = "pg13", "driver"),
  verbose = TRUE
)

Arguments

dbname
Name of a local Postgres database, Default: 'athena'

port
The port on the server to connect to, Default: 5432

Value

A connection class object to the database.

Local Connection

This function combines the process of creating a connectionDetails object and making the connection in a single function call. If a connectionDetails object is required, see makeLocalConnDetails.

This function combines the process of creating a connectionDetails object and making the connection in a single function call. If a connectionDetails object is required, see makeLocalConnDetails.

See also