Open a Connection in the pg13_connection_env

Connect to a database and assign it to the pg13_connection_env. This is a method to keep mulitple database connections organized during a crosswalk project. Existing connection objects will not be overwritten If the connection object of the provided name already exists. If the pg13_connection_env does not exist in the Global Environment, it will also be created at this point beforehand.

open_conn(conn_fun, conn_name, verbose = TRUE)

Arguments

conn_fun
Function call for the connection as a character string such as "local_connect()".

conn_name
Name of the connection object in the pg13_connection_env environment

See also