Search a Table for a Value

Loop a query for a set of one or more values in a table across all the existing fields or optionally, a subset of the fields. Both the values and the table fields are ensured compatibility by 1. Converting each value in the values argument to the character class and 2. Casting each table field as varchar in the query.

searchTable(
  conn,
  schema,
  tableName,
  ...,
  values,
  case_insensitive = TRUE,
  verbose = TRUE,
  render_sql = TRUE
)

Arguments

...
(Optional) Character strings of 1 or more fields in the table to search in.

values
Vector of length 1 or greater to search for.

case_insensitive
If TRUE, both sides of the query are converted to lowercase.

See also

Examples

library(pg13)
create_test_schema <-
        function(conn) {

                if (!schema_exists(conn = conn,
                                   schema = "test_schema")) {

                        cli::cli_rule("Create 'test_schema' Schema")

                        create_schema(conn = conn,
                                      schema = "test_schema")

                }
        }

conn <- local_connect(dbname = "pg13_test")

#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: database "pg13_test" does not exist

create_test_schema(conn = conn)

#> ── Checks ──────────────────────────────────────────────────────────────────────
#> 

#> Error in check_conn(conn = conn): object 'conn' not found

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table",
            drop_existing = TRUE,
            data = data.frame(A = 1:3, B = letters[1:3]))

#> Error in brake_closed_conn(conn = conn): connection is not open

append_table(conn = conn,
             schema = "test_schema",
             table = "test_table",
             data = data.frame(A = 1:3, B = letters[1:3]))

#> Error in brake_closed_conn(conn = conn): connection is not open

append_table(conn = conn,
             schema = "test_schema",
             table = "test_table",
             data = data.frame(A = 1:3, B = rep(NA_character_, 3)))

#> Error in brake_closed_conn(conn = conn): connection is not open

test_data <- read_table(conn = conn,
                           schema = "test_schema",
                           table = "test_table")

#> [2020-12-23 20:21:35]	SQL: SELECT * FROM test_schema.test_table;
#> [2020-12-23 20:21:35]	Reading...
#> ── Checks ──────────────────────────────────────────────────────────────────────
#> 

#> Error in check_conn(conn = conn): object 'conn' not found

test_data

#> Error in eval(expr, envir, enclos): object 'test_data' not found

search_table(conn = conn,
             schema = "test_schema",
             table = "test_table",
             values = 1:3)

#> Error in brake_closed_conn(conn = conn): connection is not open

search_table(conn = conn,
             schema = "test_schema",
             table = "test_table",
             values = "a")

#> Error in brake_closed_conn(conn = conn): connection is not open

search_table(conn = conn,
             schema = "test_schema",
             table = "test_table",
             values = c("A", "b", "C"),
             case_insensitive = FALSE)

#> Error in brake_closed_conn(conn = conn): connection is not open

drop_table(conn = conn,
           schema = "test_schema",
           table = "test_table",
           if_exists = FALSE,
           data = data.frame(A = 1:3, B = rep(NA_character_, 3)))

#> Error in brake_closed_conn(conn = conn): connection is not open

drop_schema(conn = conn,
            schema = "test_schema",
            cascade = TRUE)

#> ── Checks ──────────────────────────────────────────────────────────────────────
#> 

#> Error in check_conn(conn = conn): object 'conn' not found

dc(conn = conn)

#> [2020-12-23 20:21:35]	Postgres connection was already closed