Summarize all Fields in a Schema

summarize_field gets the total rows, distinct value count, valueset, and null counts for a single field in a table. To get the summary of more than 1 field, see summarize_fields. To summarize an entire table or schema without hand-selecting fields, see summarize_table and summarize_schema respectively.

summarize_schema(
  conn,
  conn_fun,
  schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  warn_no_rows = TRUE,
  ...
)

Examples

library(pg13)
create_test_schema <-
        function(conn) {

                if (!schema_exists(conn = conn,
                                   schema = "test_schema")) {

                        cli::cli_rule("Create 'test_schema' Schema")

                        create_schema(conn = conn,
                                      schema = "test_schema")

                }
        }

conn <- local_connect(dbname = "pg13_test")

#> Connecting using PostgreSQL driver

create_test_schema(conn = conn)

#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table",
            drop_existing = TRUE,
            data = data.frame(A = 1:3, B = letters[1:3]))

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table;
#> [2020-12-20 17:45:25]	Dropping...
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table;
#> [2020-12-20 17:45:25]	Sending...
#> [2020-12-20 17:45:25]	Sending...complete
#> [2020-12-20 17:45:25]	Dropping...complete
#> [2020-12-20 17:45:25]	SQL: N/A
#> [2020-12-20 17:45:25]	Writing...
#> [2020-12-20 17:45:25]	Writing...complete

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table2",
            drop_existing = TRUE,
            data = data.frame(C = rep(NA, 3), D = c(TRUE, FALSE, FALSE)))

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table2;
#> [2020-12-20 17:45:25]	Dropping...
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table2;
#> [2020-12-20 17:45:25]	Sending...
#> [2020-12-20 17:45:25]	Sending...complete
#> [2020-12-20 17:45:25]	Dropping...complete
#> [2020-12-20 17:45:25]	SQL: N/A
#> [2020-12-20 17:45:25]	Writing...

#> Warning: no non-missing arguments to max; returning -Inf

#> [2020-12-20 17:45:25]	Writing...complete

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table3",
            drop_existing = TRUE,
            data = data.frame(E = c(1.25, 343.31341, 5),
                              G = c(Sys.Date(), Sys.Date()-100, Sys.Date()-1000)))

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table3;
#> [2020-12-20 17:45:25]	Dropping...
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: DROP TABLE IF EXISTS test_schema.test_table3;
#> [2020-12-20 17:45:25]	Sending...
#> [2020-12-20 17:45:25]	Sending...complete
#> [2020-12-20 17:45:25]	Dropping...complete
#> [2020-12-20 17:45:25]	SQL: N/A
#> [2020-12-20 17:45:25]	Writing...
#> [2020-12-20 17:45:25]	Writing...complete

summarize_fields(conn = conn, schema = "test_schema", table = "test_table", fields = c("A", "B"))

#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT COUNT(*) FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: WITH valueset AS (SELECT DISTINCT a FROM test_schema.test_table) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT DISTINCT a FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT COUNT(a) FROM test_schema.test_table WHERE a IS NULL;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT COUNT(*) FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: WITH valueset AS (SELECT DISTINCT b FROM test_schema.test_table) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT DISTINCT b FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT COUNT(b) FROM test_schema.test_table WHERE b IS NULL;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete

#> $A
#> $A$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $A$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $A$VALUESET
#>   a
#> 1 3
#> 2 2
#> 3 1
#> 
#> $A$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $B
#> $B$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $B$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $B$VALUESET
#>   b
#> 1 a
#> 2 b
#> 3 c
#> 
#> $B$NULL_COUNT
#>   count
#> 1     0
#> 
#> 

# Case is ignored
summarize_fields(conn = conn, schema = "test_schema", table = "test_table", fields = c("a", "b"))

#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT COUNT(*) FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: WITH valueset AS (SELECT DISTINCT a FROM test_schema.test_table) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:25]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:25]	SQL: SELECT DISTINCT a FROM test_schema.test_table;
#> [2020-12-20 17:45:25]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(a) FROM test_schema.test_table WHERE a IS NULL;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(*) FROM test_schema.test_table;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: WITH valueset AS (SELECT DISTINCT b FROM test_schema.test_table) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT DISTINCT b FROM test_schema.test_table;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(b) FROM test_schema.test_table WHERE b IS NULL;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete

#> $a
#> $a$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $a$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $a$VALUESET
#>   a
#> 1 3
#> 2 2
#> 3 1
#> 
#> $a$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $b
#> $b$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $b$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $b$VALUESET
#>   b
#> 1 a
#> 2 b
#> 3 c
#> 
#> $b$NULL_COUNT
#>   count
#> 1     0
#> 
#> 

summarize_fields(conn = conn, schema = "test_schema", table = "test_table2", fields = c("c", "D"))

#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(*) FROM test_schema.test_table2;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: WITH valueset AS (SELECT DISTINCT c FROM test_schema.test_table2) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT DISTINCT c FROM test_schema.test_table2;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(c) FROM test_schema.test_table2 WHERE c IS NULL;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(*) FROM test_schema.test_table2;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: WITH valueset AS (SELECT DISTINCT d FROM test_schema.test_table2) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT DISTINCT d FROM test_schema.test_table2;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(d) FROM test_schema.test_table2 WHERE d IS NULL;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete

#> $c
#> $c$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $c$DISTINCT_VALUE_COUNT
#>   count
#> 1     1
#> 
#> $c$VALUESET
#>      c
#> 1 <NA>
#> 
#> $c$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $D
#> $D$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $D$DISTINCT_VALUE_COUNT
#>   count
#> 1     2
#> 
#> $D$VALUESET
#>       d
#> 1  TRUE
#> 2 FALSE
#> 
#> $D$NULL_COUNT
#>   count
#> 1     0
#> 
#> 

# To summarize an entire table (all the fields without manually inputting them)
summarize_table(conn = conn, schema = "test_schema", table = "test_table3")

#> [2020-12-20 17:45:26]	SQL: N/A
#> [2020-12-20 17:45:26]	Listing Fields...
#> [2020-12-20 17:45:26]	Listing Fields...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(*) FROM test_schema.test_table3;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: WITH valueset AS (SELECT DISTINCT e FROM test_schema.test_table3) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT DISTINCT e FROM test_schema.test_table3;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(e) FROM test_schema.test_table3 WHERE e IS NULL;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:26]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:26]	SQL: SELECT COUNT(*) FROM test_schema.test_table3;
#> [2020-12-20 17:45:26]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: WITH valueset AS (SELECT DISTINCT g FROM test_schema.test_table3) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT DISTINCT g FROM test_schema.test_table3;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(g) FROM test_schema.test_table3 WHERE g IS NULL;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete

#> $e
#> $e$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $e$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $e$VALUESET
#>          e
#> 1 343.3134
#> 2   1.2500
#> 3   5.0000
#> 
#> $e$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $g
#> $g$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $g$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $g$VALUESET
#>            g
#> 1 2018-03-26
#> 2 2020-12-20
#> 3 2020-09-11
#> 
#> $g$NULL_COUNT
#>   count
#> 1     0
#> 
#> 

# An entire schema can also be summarized
summarize_schema(conn = conn, schema = "test_schema")

#> [2020-12-20 17:45:27]	SQL: N/A
#> [2020-12-20 17:45:27]	Listing Tables...
#> [2020-12-20 17:45:27]	Listing Tables...completed
#> [2020-12-20 17:45:27]	SQL: N/A
#> [2020-12-20 17:45:27]	Listing Fields...
#> [2020-12-20 17:45:27]	Listing Fields...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: WITH valueset AS (SELECT DISTINCT a FROM test_schema.TEST_TABLE) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT DISTINCT a FROM test_schema.TEST_TABLE;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(a) FROM test_schema.TEST_TABLE WHERE a IS NULL;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: WITH valueset AS (SELECT DISTINCT b FROM test_schema.TEST_TABLE) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT DISTINCT b FROM test_schema.TEST_TABLE;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(b) FROM test_schema.TEST_TABLE WHERE b IS NULL;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> [2020-12-20 17:45:27]	SQL: N/A
#> [2020-12-20 17:45:27]	Listing Fields...
#> [2020-12-20 17:45:27]	Listing Fields...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE2;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: WITH valueset AS (SELECT DISTINCT c FROM test_schema.TEST_TABLE2) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT DISTINCT c FROM test_schema.TEST_TABLE2;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(c) FROM test_schema.TEST_TABLE2 WHERE c IS NULL;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE2;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: WITH valueset AS (SELECT DISTINCT d FROM test_schema.TEST_TABLE2) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT DISTINCT d FROM test_schema.TEST_TABLE2;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(d) FROM test_schema.TEST_TABLE2 WHERE d IS NULL;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:27]	Querying...complete
#> [2020-12-20 17:45:27]	SQL: N/A
#> [2020-12-20 17:45:27]	Listing Fields...
#> [2020-12-20 17:45:27]	Listing Fields...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:27]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE3;
#> [2020-12-20 17:45:27]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: WITH valueset AS (SELECT DISTINCT e FROM test_schema.TEST_TABLE3) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: SELECT DISTINCT e FROM test_schema.TEST_TABLE3;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: SELECT COUNT(e) FROM test_schema.TEST_TABLE3 WHERE e IS NULL;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: SELECT COUNT(*) FROM test_schema.TEST_TABLE3;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: WITH valueset AS (SELECT DISTINCT g FROM test_schema.TEST_TABLE3) SELECT COUNT(*) FROM valueset;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: SELECT DISTINCT g FROM test_schema.TEST_TABLE3;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete
#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: SELECT COUNT(g) FROM test_schema.TEST_TABLE3 WHERE g IS NULL;
#> [2020-12-20 17:45:28]	Querying...
#> [2020-12-20 17:45:28]	Querying...complete

#> $TEST_TABLE
#> $TEST_TABLE$a
#> $TEST_TABLE$a$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE$a$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $TEST_TABLE$a$VALUESET
#>   a
#> 1 3
#> 2 2
#> 3 1
#> 
#> $TEST_TABLE$a$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $TEST_TABLE$b
#> $TEST_TABLE$b$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE$b$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $TEST_TABLE$b$VALUESET
#>   b
#> 1 a
#> 2 b
#> 3 c
#> 
#> $TEST_TABLE$b$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> 
#> $TEST_TABLE2
#> $TEST_TABLE2$c
#> $TEST_TABLE2$c$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE2$c$DISTINCT_VALUE_COUNT
#>   count
#> 1     1
#> 
#> $TEST_TABLE2$c$VALUESET
#>      c
#> 1 <NA>
#> 
#> $TEST_TABLE2$c$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $TEST_TABLE2$d
#> $TEST_TABLE2$d$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE2$d$DISTINCT_VALUE_COUNT
#>   count
#> 1     2
#> 
#> $TEST_TABLE2$d$VALUESET
#>       d
#> 1  TRUE
#> 2 FALSE
#> 
#> $TEST_TABLE2$d$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> 
#> $TEST_TABLE3
#> $TEST_TABLE3$e
#> $TEST_TABLE3$e$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE3$e$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $TEST_TABLE3$e$VALUESET
#>          e
#> 1 343.3134
#> 2   1.2500
#> 3   5.0000
#> 
#> $TEST_TABLE3$e$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> $TEST_TABLE3$g
#> $TEST_TABLE3$g$TOTAL_ROWS
#>   count
#> 1     3
#> 
#> $TEST_TABLE3$g$DISTINCT_VALUE_COUNT
#>   count
#> 1     3
#> 
#> $TEST_TABLE3$g$VALUESET
#>            g
#> 1 2018-03-26
#> 2 2020-12-20
#> 3 2020-09-11
#> 
#> $TEST_TABLE3$g$NULL_COUNT
#>   count
#> 1     0
#> 
#> 
#> 

drop_schema(conn = conn,
            schema = "test_schema",
            cascade = TRUE)

#> Warning: 'drop_schema' is deprecated.
#> Use 'dropCascade' instead.
#> See help("Deprecated")

#> ── Checks ──────────────────────────────────────────────────────────────────────────────────────────────────────────
#> 

#> ✓ Open connection

#> [2020-12-20 17:45:28]	SQL: DROP SCHEMA IF EXISTS test_schema ;
#> [2020-12-20 17:45:28]	Sending...

#> Error in rJava::.jcall(statement, "V", "close"): org.postgresql.util.PSQLException: ERROR: cannot drop schema test_schema because other objects depend on it
#>   Detail: table test_schema.test_table depends on schema test_schema
#> table test_schema.test_table2 depends on schema test_schema
#> table test_schema.test_table3 depends on schema test_schema
#>   Hint: Use DROP ... CASCADE to drop the dependent objects too.

dc(conn = conn)

#> [2020-12-20 17:45:28]	Postgres connection closed