Write a Staging Table.

A Staging Table is one that is named automatically in a naming convention of "V" followed by 14 integers representing the timestamp of the transaction. The staging table can optionally be dropped on exit in the parent frame from which the function is being called when drop_on_exit is set to TRUE.

write_staging_table(
  conn,
  conn_fun,
  schema,
  data,
  drop_existing = FALSE,
  drop_on_exit = TRUE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)

Arguments

conn
Connection object

schema
The target schema for the operation.

data
A dataframe or tibble.

verbose
If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.

render_sql
If TRUE, the SQL statement for the operation is returned in the console.

Value

String of the staging table

See also

Examples

library(pg13)
library(tidyverse)
test_data <-
        tibble::tibble(A = 1:3,
                       B = letters[1:3],
                       C = c(TRUE, FALSE, TRUE),
                       D = c(1.23421, 234.23421, 342.0134014134),
                       E = c(Sys.Date(), Sys.Date()-1013, Sys.Date() + 134))
conn <- local_connect("pg13_test")

#> Connecting using PostgreSQL driver

# Writing a staging table. Note that since the function is not called
# inside a enclosed parent frame, the table is not dropped even when
# drop_on_exit is set to TRUE
write_staging_table(conn = conn,
                    schema = "public",
                    data = test_data,
                    drop_on_exit = TRUE)

#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2020-12-25 17:52:18] 

#> ✓ Table name 'V20201225175218' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'a' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'b' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'c' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'd' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'e' is not a reserved word

#> [2020-12-25 17:52:18]	Writing public.V20201225175218...
#> [2020-12-25 17:52:18]	Writing public.V20201225175218...complete
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...
#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18]	SQL: DROP TABLE IF EXISTS public.V20201225175218;
#> [2020-12-25 17:52:18]	Sending...
#> [2020-12-25 17:52:18]	Sending...complete
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...complete

#> [1] "V20201225175218"

ls_tables(conn = conn,
          schema = "public")

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed

#> character(0)

# All tables in a schema can be dropped with pattern matching to the "V{timestamp}" format
drop_all_staging_tables(conn = conn,
                        schema = "public")

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed
#> [2020-12-25 17:52:18]	0 tables in public

# The `time_diff_hours` option allows for the user to drop all tables other than the most recently written tables based on the number of hours from the current timestamp.
write_staging_table(conn = conn,
                    schema = "public",
                    data = test_data)

#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2020-12-25 17:52:18] 

#> ✓ Table name 'V20201225175218' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'a' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'b' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'c' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'd' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'e' is not a reserved word

#> [2020-12-25 17:52:18]	Writing public.V20201225175218...
#> [2020-12-25 17:52:18]	Writing public.V20201225175218...complete
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...
#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18]	SQL: DROP TABLE IF EXISTS public.V20201225175218;
#> [2020-12-25 17:52:18]	Sending...
#> [2020-12-25 17:52:18]	Sending...complete
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...complete

#> [1] "V20201225175218"

drop_all_staging_tables(conn = conn,
                        schema = "public",
                        time_diff_hours = 8)

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed
#> [2020-12-25 17:52:18]	0 tables in public

ls_tables(conn = conn,
          schema = "public")

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed

#> character(0)

# If this function is called within an enclosed parent frame with `drop_on_exit` set to TRUE, the table will be dropped at the conclusion of the function.
test_fun <-
        function() {

                new_table <-
                write_staging_table(conn = conn,
                                    schema = "public",
                                    data = test_data,
                                    drop_on_exit = TRUE)

                print("The test is finished.")
        }

test_fun()

#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2020-12-25 17:52:18] 

#> ✓ Table name 'V20201225175218' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'a' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'b' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'c' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'd' is not a reserved word

#> [2020-12-25 17:52:18] 

#> ✓ Field name 'e' is not a reserved word

#> [2020-12-25 17:52:18]	Writing public.V20201225175218...
#> [2020-12-25 17:52:18]	Writing public.V20201225175218...complete
#> [1] "The test is finished."
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...
#> [2020-12-25 17:52:18] 

#> ✓ Open connection

#> [2020-12-25 17:52:18] 

#> ✓ JDBC connection

#> [2020-12-25 17:52:18]	SQL: DROP TABLE IF EXISTS public.V20201225175218;
#> [2020-12-25 17:52:18]	Sending...
#> [2020-12-25 17:52:18]	Sending...complete
#> [2020-12-25 17:52:18]	Dropping public.V20201225175218...complete

ls_tables(conn = conn,
          schema = "public")

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed

#> character(0)

# Dropping all staging tables before exiting the example
drop_all_staging_tables(conn = conn,
                        schema = "public",
                        time_diff_hours = 8)

#> [2020-12-25 17:52:18]	SQL: N/A
#> [2020-12-25 17:52:18]	Listing Tables...
#> [2020-12-25 17:52:18]	Listing Tables...completed
#> [2020-12-25 17:52:18]	0 tables in public

dc(conn = conn)

#> [2020-12-25 17:52:18]	Postgres connection to 'pg13_test' closed