Search a Table for a Value

Loop a query for a set of one or more values in a table across all the existing fields or optionally, a subset of the fields. Both the values and the table fields are ensured compatibility by 1. Converting each value in the values argument to the character class and 2. Casting each table field as varchar in the query.

searchTable(
  conn,
  schema,
  tableName,
  ...,
  values,
  case_insensitive = TRUE,
  verbose = TRUE,
  render_sql = TRUE
)

Arguments

conn
A connection object returned by connect().

schema
The target schema for the operation.

tableName
The target table in the schema for an operation or for writing functions, the name of the table that will be created.

...
(Optional) Character strings of 1 or more fields in the table to search in.

values
Vector of length 1 or greater to search for.

case_insensitive
If TRUE, both sides of the query are converted to lowercase.

verbose
If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.

render_sql
If TRUE, the SQL statement for the operation is returned in the console.

See also

Examples

conn <- localConnect(dbname = "athena",
                     port = 5432)

#> Connecting using PostgreSQL driver

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = TRUE)

#> [2020-11-29 18:43:28]	Listing Fields...
#> [2020-11-29 18:43:28]	Listing Fields...complete
#> [2020-11-29 18:43:28]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.concept_id::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:28]	Querying...
#> [2020-11-29 18:43:33]	Querying...complete
#> [2020-11-29 18:43:33]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.concept_name::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:33]	Querying...
#> [2020-11-29 18:43:40]	Querying...complete
#> [2020-11-29 18:43:40]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.domain_id::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:40]	Querying...
#> [2020-11-29 18:43:45]	Querying...complete
#> [2020-11-29 18:43:45]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.vocabulary_id::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:45]	Querying...
#> [2020-11-29 18:43:49]	Querying...complete
#> [2020-11-29 18:43:49]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.concept_class_id::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:49]	Querying...
#> [2020-11-29 18:43:54]	Querying...complete
#> [2020-11-29 18:43:54]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.standard_concept::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:54]	Querying...
#> [2020-11-29 18:43:56]	Querying...complete
#> [2020-11-29 18:43:56]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.concept_code::varchar) IN ('aspirin') ;
#> [2020-11-29 18:43:56]	Querying...
#> [2020-11-29 18:44:01]	Querying...complete
#> [2020-11-29 18:44:01]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.valid_start_date::varchar) IN ('aspirin') ;
#> [2020-11-29 18:44:01]	Querying...
#> [2020-11-29 18:44:07]	Querying...complete
#> [2020-11-29 18:44:07]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.valid_end_date::varchar) IN ('aspirin') ;
#> [2020-11-29 18:44:07]	Querying...
#> [2020-11-29 18:44:14]	Querying...complete
#> [2020-11-29 18:44:14]	SQL: SELECT * FROM public.concept t WHERE LOWER(t.invalid_reason::varchar) IN ('aspirin') ;
#> [2020-11-29 18:44:14]	Querying...
#> [2020-11-29 18:44:16]	Querying...complete

#> $ROWS
#> # A tibble: 10 x 2
#>    Field             Rows
#>    <chr>            <int>
#>  1 concept_id           0
#>  2 concept_name        51
#>  3 domain_id            0
#>  4 vocabulary_id        0
#>  5 concept_class_id     0
#>  6 standard_concept     0
#>  7 concept_code         0
#>  8 valid_start_date     0
#>  9 valid_end_date       0
#> 10 invalid_reason       0
#> 
#> $RESULTSETS
#> $RESULTSETS$concept_name
#>    concept_id concept_name   domain_id    vocabulary_id    concept_class_id
#> 1     1112807      aspirin        Drug           RxNorm          Ingredient
#> 2     4194196      Aspirin        Drug           SNOMED Pharma/Biol Product
#> 3     4287142      ASPIRIN        Drug            NDFRT  Pharma Preparation
#> 4     4306886      Aspirin        Drug           SNOMED           Substance
#> 5     4325480      Aspirin        Drug            NDFRT  Chemical Structure
#> 6    21197519      Aspirin        Drug             dm+d          Ingredient
#> 7    21234362      Aspirin        Drug             dm+d          Ingredient
#> 8    21272111      Aspirin        Drug             dm+d                 VTM
#> 9    35144833      Aspirin        Drug             JMDC        Drug Product
#> 10   35145382      Aspirin        Drug             JMDC        Drug Product
#> 11   35164836      aspirin        Drug             JMDC          Ingredient
#> 12   35802900      Aspirin        Drug           HemOnc           Component
#> 13   36308940      Aspirin  Meas Value            LOINC              Answer
#> 14   37491813      Aspirin        Drug              CTD          Ingredient
#> 15   43368904      Aspirin        Drug              AMT        AU Substance
#> 16   43565741      Aspirin        Drug RxNorm Extension          Brand Name
#> 17   45615296      Aspirin        Drug             MeSH        Main Heading
#> 18   45623494      aspirin        Drug              SPL                 SPL
#> 19   45627474      Aspirin        Drug              SPL                 SPL
#> 20   45633182      Aspirin        Drug              SPL                 SPL
#> 21   45639008      Aspirin        Drug              SPL                 SPL
#> 22   45639297      Aspirin        Drug              SPL                 SPL
#> 23   45639553      Aspirin        Drug              SPL                 SPL
#> 24   45640614      ASPIRIN        Drug              SPL                 SPL
#> 25   45643575      Aspirin        Drug              SPL                 SPL
#> 26   45645704      Aspirin        Drug              SPL                 SPL
#> 27   45649656      Aspirin        Drug              SPL                 SPL
#> 28   45652097      Aspirin        Drug              SPL                 SPL
#> 29   45654782      aspirin        Drug              SPL                 SPL
#> 30   45661387      Aspirin        Drug              SPL                 SPL
#> 31   45662675      Aspirin        Drug              SPL                 SPL
#> 32   45662754      Aspirin        Drug              SPL                 SPL
#> 33   45663354      Aspirin        Drug              SPL                 SPL
#> 34   45665235      Aspirin        Drug              SPL                 SPL
#> 35   45671751      Aspirin        Drug              SPL                 SPL
#> 36   45672031      Aspirin        Drug              SPL                 SPL
#> 37   45673209      Aspirin        Drug              SPL                 SPL
#> 38   45674948      aspirin        Drug              SPL                 SPL
#> 39   45678154      Aspirin        Drug              SPL                 SPL
#> 40   45685766      Aspirin        Drug              SPL                 SPL
#> 41   45687141      aspirin        Drug              SPL                 SPL
#> 42   45687425      Aspirin        Drug              SPL                 SPL
#> 43   45689848      Aspirin        Drug              SPL                 SPL
#> 44   45690793      Aspirin        Drug              SPL                 SPL
#> 45   45692943      Aspirin        Drug              SPL                 SPL
#> 46   45693677      Aspirin        Drug              SPL                 SPL
#> 47   45693813      Aspirin        Drug              SPL                 SPL
#> 48   45699321      Aspirin        Drug              SPL                 SPL
#> 49   45703548      ASPIRIN        Drug              SPL                 SPL
#> 50   45947071      ASPIRIN        Drug             CIEL                Drug
#> 51   45957786      Aspirin Observation           SNOMED           Substance
#>    standard_concept                         concept_code valid_start_date
#> 1                 S                                 1191       1970-01-01
#> 2              <NA>                              7947003       1970-01-01
#> 3              <NA>                          N0000145918       1970-01-01
#> 4              <NA>                            387458008       2005-07-27
#> 5              <NA>                          N0000006582       1970-01-01
#> 6              <NA>                     3587011000001109       1970-01-01
#> 7              <NA>                            387458008       2005-07-27
#> 8              <NA>                              7947003       1970-01-01
#> 9              <NA>                         100000042000       1970-01-01
#> 10             <NA>                         100000071579       1970-01-01
#> 11             <NA>                             JMDC1103       1970-01-01
#> 12             <NA>                                   47       2019-05-27
#> 13                S                            LA26702-3       1970-01-01
#> 14             <NA>                              50-78-2       1970-01-01
#> 15             <NA>                     1817011000036100       2016-11-01
#> 16             <NA>                           OMOP883993       1970-01-01
#> 17             <NA>                              D001241       1970-01-01
#> 18             <NA> b502e991-d503-4d8c-8626-26ecea3abc4e       1970-01-01
#> 19             <NA> 6c5da7fb-6e7b-4ac3-8a75-728c1e50fe1d       1970-01-01
#> 20             <NA> 7136e21e-49ad-4b7a-a9a1-7cfe3f1d80d5       1970-01-01
#> 21             <NA> 8174dd75-2e11-4d4f-aa69-2b3a33a19b0d       1970-01-01
#> 22             <NA> 90b5a964-00cb-4a1c-a19a-442bfa0f18bf       1970-01-01
#> 23             <NA> 9c0a0b9b-ef6f-4a60-acca-84dc815deae5       1970-01-01
#> 24             <NA> cb922a05-8937-4e3a-94fa-cda17a4ae3d2       1970-01-01
#> 25             <NA> 5264ec86-1571-4d7a-a5d9-a52b2e4fcdf3       1970-01-01
#> 26             <NA> b528bc43-d350-4d03-a463-720e52532749       1970-01-01
#> 27             <NA> 6c584de6-2131-4d48-a3f5-18408aff9b2d       1970-01-01
#> 28             <NA> daeaa324-6546-4330-a9d3-b0b93da2d853       1970-01-01
#> 29             <NA> 5387438b-1e16-43f7-915f-21adbce6af88       1970-01-01
#> 30             <NA> 87e43d62-aaa7-459f-8202-a629a4d7a423       1970-01-01
#> 31             <NA> c4efae47-dfb4-4102-a620-64631daef629       1970-01-01
#> 32             <NA> c86cffe1-949b-48b2-b5fb-388c6579de84       1970-01-01
#> 33             <NA> e4c26347-6af0-403a-81f6-4dbef95d5553       1970-01-01
#> 34             <NA> 3b8a0939-9e08-46da-8b3c-fba9c3fea9ff       1970-01-01
#> 35             <NA> 6a82cf66-a806-4331-b6a6-d13a85ec6fec       1970-01-01
#> 36             <NA> 76f12351-1855-4901-a042-b919a3539787       1970-01-01
#> 37             <NA> ac4fe306-e45d-4855-98c9-eca227b51f92       1970-01-01
#> 38             <NA> fbbf5001-c25f-4985-921d-ffa42a8e02d0       1970-01-01
#> 39             <NA> 8e161987-6226-41b7-82ef-1536ad7b1251       1970-01-01
#> 40             <NA> ee276b52-7c66-45c4-9c88-c10f47f4dbb0       1970-01-01
#> 41             <NA> 2d54e2ec-fbec-43ba-9401-085c6329ebdc       1970-01-01
#> 42             <NA> 3a650603-ebcd-4239-83d3-0a4447bd288c       1970-01-01
#> 43             <NA> a9dfc258-6e3d-48fb-a6f2-b98d6f7cea47       1970-01-01
#> 44             <NA> d45d19b1-38cd-408c-a1e4-e01e7d2014ff       1970-01-01
#> 45             <NA> 37af67f9-9741-4d67-8eb6-f6f29a60f26b       1970-01-01
#> 46             <NA> 59584a64-3b7f-4180-86eb-cde242458f1c       1970-01-01
#> 47             <NA> 601c6c97-3141-492d-869c-024fafee1783       1970-01-01
#> 48             <NA> 5ee6ca45-445f-4219-8618-e76ed2e5a70d       1970-01-01
#> 49             <NA> 1f3d7d96-1aad-4d0c-bc4a-c8960a0c683a       1970-01-01
#> 50             <NA>                                   88       2004-01-01
#> 51             <NA>                     3587011000001109       1970-01-01
#>    valid_end_date invalid_reason
#> 1      2099-12-31           <NA>
#> 2      2099-12-31           <NA>
#> 3      2099-12-31           <NA>
#> 4      2099-12-31           <NA>
#> 5      2099-12-31           <NA>
#> 6      2005-07-26              D
#> 7      2099-12-31           <NA>
#> 8      2018-06-03              D
#> 9      2099-12-31           <NA>
#> 10     2099-12-31           <NA>
#> 11     2099-12-31           <NA>
#> 12     2099-12-31           <NA>
#> 13     2099-12-31           <NA>
#> 14     2099-12-31           <NA>
#> 15     2099-12-31           <NA>
#> 16     2017-08-23              D
#> 17     2099-12-31           <NA>
#> 18     2099-12-31           <NA>
#> 19     2099-12-31           <NA>
#> 20     2099-12-31           <NA>
#> 21     2099-12-31           <NA>
#> 22     2099-12-31           <NA>
#> 23     2099-12-31           <NA>
#> 24     2099-12-31           <NA>
#> 25     2099-12-31           <NA>
#> 26     2099-12-31           <NA>
#> 27     2099-12-31           <NA>
#> 28     2099-12-31           <NA>
#> 29     2099-12-31           <NA>
#> 30     2099-12-31           <NA>
#> 31     2099-12-31           <NA>
#> 32     2099-12-31           <NA>
#> 33     2099-12-31           <NA>
#> 34     2099-12-31           <NA>
#> 35     2099-12-31           <NA>
#> 36     2099-12-31           <NA>
#> 37     2099-12-31           <NA>
#> 38     2099-12-31           <NA>
#> 39     2099-12-31           <NA>
#> 40     2099-12-31           <NA>
#> 41     2099-12-31           <NA>
#> 42     2099-12-31           <NA>
#> 43     2099-12-31           <NA>
#> 44     2099-12-31           <NA>
#> 45     2099-12-31           <NA>
#> 46     2099-12-31           <NA>
#> 47     2099-12-31           <NA>
#> 48     2099-12-31           <NA>
#> 49     2099-12-31           <NA>
#> 50     2015-02-27              D
#> 51     2005-07-26              U
#> 
#> 

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE)

#> [2020-11-29 18:44:16]	Listing Fields...
#> [2020-11-29 18:44:16]	Listing Fields...complete
#> [2020-11-29 18:44:17]	SQL: SELECT * FROM public.concept t WHERE t.concept_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:17]	Querying...
#> [2020-11-29 18:44:17]	Querying...complete
#> [2020-11-29 18:44:17]	SQL: SELECT * FROM public.concept t WHERE t.concept_name::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:17]	Querying...
#> [2020-11-29 18:44:18]	Querying...complete
#> [2020-11-29 18:44:18]	SQL: SELECT * FROM public.concept t WHERE t.domain_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:18]	Querying...
#> [2020-11-29 18:44:18]	Querying...complete
#> [2020-11-29 18:44:18]	SQL: SELECT * FROM public.concept t WHERE t.vocabulary_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:18]	Querying...
#> [2020-11-29 18:44:18]	Querying...complete
#> [2020-11-29 18:44:18]	SQL: SELECT * FROM public.concept t WHERE t.concept_class_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:18]	Querying...
#> [2020-11-29 18:44:18]	Querying...complete
#> [2020-11-29 18:44:18]	SQL: SELECT * FROM public.concept t WHERE t.standard_concept::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:18]	Querying...
#> [2020-11-29 18:44:20]	Querying...complete
#> [2020-11-29 18:44:20]	SQL: SELECT * FROM public.concept t WHERE t.concept_code::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:20]	Querying...
#> [2020-11-29 18:44:20]	Querying...complete
#> [2020-11-29 18:44:20]	SQL: SELECT * FROM public.concept t WHERE t.valid_start_date::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:20]	Querying...
#> [2020-11-29 18:44:23]	Querying...complete
#> [2020-11-29 18:44:23]	SQL: SELECT * FROM public.concept t WHERE t.valid_end_date::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:23]	Querying...
#> [2020-11-29 18:44:27]	Querying...complete
#> [2020-11-29 18:44:27]	SQL: SELECT * FROM public.concept t WHERE t.invalid_reason::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:27]	Querying...
#> [2020-11-29 18:44:30]	Querying...complete

#> $ROWS
#> # A tibble: 10 x 2
#>    Field             Rows
#>    <chr>            <int>
#>  1 concept_id           0
#>  2 concept_name        41
#>  3 domain_id            0
#>  4 vocabulary_id        0
#>  5 concept_class_id     0
#>  6 standard_concept     0
#>  7 concept_code         0
#>  8 valid_start_date     0
#>  9 valid_end_date       0
#> 10 invalid_reason       0
#> 
#> $RESULTSETS
#> $RESULTSETS$concept_name
#>    concept_id concept_name   domain_id    vocabulary_id    concept_class_id
#> 1     4194196      Aspirin        Drug           SNOMED Pharma/Biol Product
#> 2     4306886      Aspirin        Drug           SNOMED           Substance
#> 3     4325480      Aspirin        Drug            NDFRT  Chemical Structure
#> 4    21197519      Aspirin        Drug             dm+d          Ingredient
#> 5    21234362      Aspirin        Drug             dm+d          Ingredient
#> 6    21272111      Aspirin        Drug             dm+d                 VTM
#> 7    35144833      Aspirin        Drug             JMDC        Drug Product
#> 8    35145382      Aspirin        Drug             JMDC        Drug Product
#> 9    35802900      Aspirin        Drug           HemOnc           Component
#> 10   36308940      Aspirin  Meas Value            LOINC              Answer
#> 11   37491813      Aspirin        Drug              CTD          Ingredient
#> 12   43368904      Aspirin        Drug              AMT        AU Substance
#> 13   43565741      Aspirin        Drug RxNorm Extension          Brand Name
#> 14   45615296      Aspirin        Drug             MeSH        Main Heading
#> 15   45627474      Aspirin        Drug              SPL                 SPL
#> 16   45633182      Aspirin        Drug              SPL                 SPL
#> 17   45639008      Aspirin        Drug              SPL                 SPL
#> 18   45639297      Aspirin        Drug              SPL                 SPL
#> 19   45639553      Aspirin        Drug              SPL                 SPL
#> 20   45643575      Aspirin        Drug              SPL                 SPL
#> 21   45645704      Aspirin        Drug              SPL                 SPL
#> 22   45649656      Aspirin        Drug              SPL                 SPL
#> 23   45652097      Aspirin        Drug              SPL                 SPL
#> 24   45661387      Aspirin        Drug              SPL                 SPL
#> 25   45662675      Aspirin        Drug              SPL                 SPL
#> 26   45662754      Aspirin        Drug              SPL                 SPL
#> 27   45663354      Aspirin        Drug              SPL                 SPL
#> 28   45665235      Aspirin        Drug              SPL                 SPL
#> 29   45671751      Aspirin        Drug              SPL                 SPL
#> 30   45672031      Aspirin        Drug              SPL                 SPL
#> 31   45673209      Aspirin        Drug              SPL                 SPL
#> 32   45678154      Aspirin        Drug              SPL                 SPL
#> 33   45685766      Aspirin        Drug              SPL                 SPL
#> 34   45687425      Aspirin        Drug              SPL                 SPL
#> 35   45689848      Aspirin        Drug              SPL                 SPL
#> 36   45690793      Aspirin        Drug              SPL                 SPL
#> 37   45692943      Aspirin        Drug              SPL                 SPL
#> 38   45693677      Aspirin        Drug              SPL                 SPL
#> 39   45693813      Aspirin        Drug              SPL                 SPL
#> 40   45699321      Aspirin        Drug              SPL                 SPL
#> 41   45957786      Aspirin Observation           SNOMED           Substance
#>    standard_concept                         concept_code valid_start_date
#> 1              <NA>                              7947003       1970-01-01
#> 2              <NA>                            387458008       2005-07-27
#> 3              <NA>                          N0000006582       1970-01-01
#> 4              <NA>                     3587011000001109       1970-01-01
#> 5              <NA>                            387458008       2005-07-27
#> 6              <NA>                              7947003       1970-01-01
#> 7              <NA>                         100000042000       1970-01-01
#> 8              <NA>                         100000071579       1970-01-01
#> 9              <NA>                                   47       2019-05-27
#> 10                S                            LA26702-3       1970-01-01
#> 11             <NA>                              50-78-2       1970-01-01
#> 12             <NA>                     1817011000036100       2016-11-01
#> 13             <NA>                           OMOP883993       1970-01-01
#> 14             <NA>                              D001241       1970-01-01
#> 15             <NA> 6c5da7fb-6e7b-4ac3-8a75-728c1e50fe1d       1970-01-01
#> 16             <NA> 7136e21e-49ad-4b7a-a9a1-7cfe3f1d80d5       1970-01-01
#> 17             <NA> 8174dd75-2e11-4d4f-aa69-2b3a33a19b0d       1970-01-01
#> 18             <NA> 90b5a964-00cb-4a1c-a19a-442bfa0f18bf       1970-01-01
#> 19             <NA> 9c0a0b9b-ef6f-4a60-acca-84dc815deae5       1970-01-01
#> 20             <NA> 5264ec86-1571-4d7a-a5d9-a52b2e4fcdf3       1970-01-01
#> 21             <NA> b528bc43-d350-4d03-a463-720e52532749       1970-01-01
#> 22             <NA> 6c584de6-2131-4d48-a3f5-18408aff9b2d       1970-01-01
#> 23             <NA> daeaa324-6546-4330-a9d3-b0b93da2d853       1970-01-01
#> 24             <NA> 87e43d62-aaa7-459f-8202-a629a4d7a423       1970-01-01
#> 25             <NA> c4efae47-dfb4-4102-a620-64631daef629       1970-01-01
#> 26             <NA> c86cffe1-949b-48b2-b5fb-388c6579de84       1970-01-01
#> 27             <NA> e4c26347-6af0-403a-81f6-4dbef95d5553       1970-01-01
#> 28             <NA> 3b8a0939-9e08-46da-8b3c-fba9c3fea9ff       1970-01-01
#> 29             <NA> 6a82cf66-a806-4331-b6a6-d13a85ec6fec       1970-01-01
#> 30             <NA> 76f12351-1855-4901-a042-b919a3539787       1970-01-01
#> 31             <NA> ac4fe306-e45d-4855-98c9-eca227b51f92       1970-01-01
#> 32             <NA> 8e161987-6226-41b7-82ef-1536ad7b1251       1970-01-01
#> 33             <NA> ee276b52-7c66-45c4-9c88-c10f47f4dbb0       1970-01-01
#> 34             <NA> 3a650603-ebcd-4239-83d3-0a4447bd288c       1970-01-01
#> 35             <NA> a9dfc258-6e3d-48fb-a6f2-b98d6f7cea47       1970-01-01
#> 36             <NA> d45d19b1-38cd-408c-a1e4-e01e7d2014ff       1970-01-01
#> 37             <NA> 37af67f9-9741-4d67-8eb6-f6f29a60f26b       1970-01-01
#> 38             <NA> 59584a64-3b7f-4180-86eb-cde242458f1c       1970-01-01
#> 39             <NA> 601c6c97-3141-492d-869c-024fafee1783       1970-01-01
#> 40             <NA> 5ee6ca45-445f-4219-8618-e76ed2e5a70d       1970-01-01
#> 41             <NA>                     3587011000001109       1970-01-01
#>    valid_end_date invalid_reason
#> 1      2099-12-31           <NA>
#> 2      2099-12-31           <NA>
#> 3      2099-12-31           <NA>
#> 4      2005-07-26              D
#> 5      2099-12-31           <NA>
#> 6      2018-06-03              D
#> 7      2099-12-31           <NA>
#> 8      2099-12-31           <NA>
#> 9      2099-12-31           <NA>
#> 10     2099-12-31           <NA>
#> 11     2099-12-31           <NA>
#> 12     2099-12-31           <NA>
#> 13     2017-08-23              D
#> 14     2099-12-31           <NA>
#> 15     2099-12-31           <NA>
#> 16     2099-12-31           <NA>
#> 17     2099-12-31           <NA>
#> 18     2099-12-31           <NA>
#> 19     2099-12-31           <NA>
#> 20     2099-12-31           <NA>
#> 21     2099-12-31           <NA>
#> 22     2099-12-31           <NA>
#> 23     2099-12-31           <NA>
#> 24     2099-12-31           <NA>
#> 25     2099-12-31           <NA>
#> 26     2099-12-31           <NA>
#> 27     2099-12-31           <NA>
#> 28     2099-12-31           <NA>
#> 29     2099-12-31           <NA>
#> 30     2099-12-31           <NA>
#> 31     2099-12-31           <NA>
#> 32     2099-12-31           <NA>
#> 33     2099-12-31           <NA>
#> 34     2099-12-31           <NA>
#> 35     2099-12-31           <NA>
#> 36     2099-12-31           <NA>
#> 37     2099-12-31           <NA>
#> 38     2099-12-31           <NA>
#> 39     2099-12-31           <NA>
#> 40     2099-12-31           <NA>
#> 41     2005-07-26              U
#> 
#> 

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE)

#> [2020-11-29 18:44:30]	SQL: SELECT * FROM public.concept t WHERE t.concept_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:30]	SQL: SELECT * FROM public.concept t WHERE t.concept_name::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:31]	SQL: SELECT * FROM public.concept t WHERE t.domain_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:31]	SQL: SELECT * FROM public.concept t WHERE t.vocabulary_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:31]	SQL: SELECT * FROM public.concept t WHERE t.concept_class_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:31]	SQL: SELECT * FROM public.concept t WHERE t.standard_concept::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:33]	SQL: SELECT * FROM public.concept t WHERE t.concept_code::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:33]	SQL: SELECT * FROM public.concept t WHERE t.valid_start_date::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:37]	SQL: SELECT * FROM public.concept t WHERE t.valid_end_date::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:41]	SQL: SELECT * FROM public.concept t WHERE t.invalid_reason::varchar IN ('Aspirin') ;

#> $ROWS
#> # A tibble: 10 x 2
#>    Field             Rows
#>    <chr>            <int>
#>  1 concept_id           0
#>  2 concept_name        41
#>  3 domain_id            0
#>  4 vocabulary_id        0
#>  5 concept_class_id     0
#>  6 standard_concept     0
#>  7 concept_code         0
#>  8 valid_start_date     0
#>  9 valid_end_date       0
#> 10 invalid_reason       0
#> 
#> $RESULTSETS
#> $RESULTSETS$concept_name
#>    concept_id concept_name   domain_id    vocabulary_id    concept_class_id
#> 1     4194196      Aspirin        Drug           SNOMED Pharma/Biol Product
#> 2     4306886      Aspirin        Drug           SNOMED           Substance
#> 3     4325480      Aspirin        Drug            NDFRT  Chemical Structure
#> 4    21197519      Aspirin        Drug             dm+d          Ingredient
#> 5    21234362      Aspirin        Drug             dm+d          Ingredient
#> 6    21272111      Aspirin        Drug             dm+d                 VTM
#> 7    35144833      Aspirin        Drug             JMDC        Drug Product
#> 8    35145382      Aspirin        Drug             JMDC        Drug Product
#> 9    35802900      Aspirin        Drug           HemOnc           Component
#> 10   36308940      Aspirin  Meas Value            LOINC              Answer
#> 11   37491813      Aspirin        Drug              CTD          Ingredient
#> 12   43368904      Aspirin        Drug              AMT        AU Substance
#> 13   43565741      Aspirin        Drug RxNorm Extension          Brand Name
#> 14   45615296      Aspirin        Drug             MeSH        Main Heading
#> 15   45627474      Aspirin        Drug              SPL                 SPL
#> 16   45633182      Aspirin        Drug              SPL                 SPL
#> 17   45639008      Aspirin        Drug              SPL                 SPL
#> 18   45639297      Aspirin        Drug              SPL                 SPL
#> 19   45639553      Aspirin        Drug              SPL                 SPL
#> 20   45643575      Aspirin        Drug              SPL                 SPL
#> 21   45645704      Aspirin        Drug              SPL                 SPL
#> 22   45649656      Aspirin        Drug              SPL                 SPL
#> 23   45652097      Aspirin        Drug              SPL                 SPL
#> 24   45661387      Aspirin        Drug              SPL                 SPL
#> 25   45662675      Aspirin        Drug              SPL                 SPL
#> 26   45662754      Aspirin        Drug              SPL                 SPL
#> 27   45663354      Aspirin        Drug              SPL                 SPL
#> 28   45665235      Aspirin        Drug              SPL                 SPL
#> 29   45671751      Aspirin        Drug              SPL                 SPL
#> 30   45672031      Aspirin        Drug              SPL                 SPL
#> 31   45673209      Aspirin        Drug              SPL                 SPL
#> 32   45678154      Aspirin        Drug              SPL                 SPL
#> 33   45685766      Aspirin        Drug              SPL                 SPL
#> 34   45687425      Aspirin        Drug              SPL                 SPL
#> 35   45689848      Aspirin        Drug              SPL                 SPL
#> 36   45690793      Aspirin        Drug              SPL                 SPL
#> 37   45692943      Aspirin        Drug              SPL                 SPL
#> 38   45693677      Aspirin        Drug              SPL                 SPL
#> 39   45693813      Aspirin        Drug              SPL                 SPL
#> 40   45699321      Aspirin        Drug              SPL                 SPL
#> 41   45957786      Aspirin Observation           SNOMED           Substance
#>    standard_concept                         concept_code valid_start_date
#> 1              <NA>                              7947003       1970-01-01
#> 2              <NA>                            387458008       2005-07-27
#> 3              <NA>                          N0000006582       1970-01-01
#> 4              <NA>                     3587011000001109       1970-01-01
#> 5              <NA>                            387458008       2005-07-27
#> 6              <NA>                              7947003       1970-01-01
#> 7              <NA>                         100000042000       1970-01-01
#> 8              <NA>                         100000071579       1970-01-01
#> 9              <NA>                                   47       2019-05-27
#> 10                S                            LA26702-3       1970-01-01
#> 11             <NA>                              50-78-2       1970-01-01
#> 12             <NA>                     1817011000036100       2016-11-01
#> 13             <NA>                           OMOP883993       1970-01-01
#> 14             <NA>                              D001241       1970-01-01
#> 15             <NA> 6c5da7fb-6e7b-4ac3-8a75-728c1e50fe1d       1970-01-01
#> 16             <NA> 7136e21e-49ad-4b7a-a9a1-7cfe3f1d80d5       1970-01-01
#> 17             <NA> 8174dd75-2e11-4d4f-aa69-2b3a33a19b0d       1970-01-01
#> 18             <NA> 90b5a964-00cb-4a1c-a19a-442bfa0f18bf       1970-01-01
#> 19             <NA> 9c0a0b9b-ef6f-4a60-acca-84dc815deae5       1970-01-01
#> 20             <NA> 5264ec86-1571-4d7a-a5d9-a52b2e4fcdf3       1970-01-01
#> 21             <NA> b528bc43-d350-4d03-a463-720e52532749       1970-01-01
#> 22             <NA> 6c584de6-2131-4d48-a3f5-18408aff9b2d       1970-01-01
#> 23             <NA> daeaa324-6546-4330-a9d3-b0b93da2d853       1970-01-01
#> 24             <NA> 87e43d62-aaa7-459f-8202-a629a4d7a423       1970-01-01
#> 25             <NA> c4efae47-dfb4-4102-a620-64631daef629       1970-01-01
#> 26             <NA> c86cffe1-949b-48b2-b5fb-388c6579de84       1970-01-01
#> 27             <NA> e4c26347-6af0-403a-81f6-4dbef95d5553       1970-01-01
#> 28             <NA> 3b8a0939-9e08-46da-8b3c-fba9c3fea9ff       1970-01-01
#> 29             <NA> 6a82cf66-a806-4331-b6a6-d13a85ec6fec       1970-01-01
#> 30             <NA> 76f12351-1855-4901-a042-b919a3539787       1970-01-01
#> 31             <NA> ac4fe306-e45d-4855-98c9-eca227b51f92       1970-01-01
#> 32             <NA> 8e161987-6226-41b7-82ef-1536ad7b1251       1970-01-01
#> 33             <NA> ee276b52-7c66-45c4-9c88-c10f47f4dbb0       1970-01-01
#> 34             <NA> 3a650603-ebcd-4239-83d3-0a4447bd288c       1970-01-01
#> 35             <NA> a9dfc258-6e3d-48fb-a6f2-b98d6f7cea47       1970-01-01
#> 36             <NA> d45d19b1-38cd-408c-a1e4-e01e7d2014ff       1970-01-01
#> 37             <NA> 37af67f9-9741-4d67-8eb6-f6f29a60f26b       1970-01-01
#> 38             <NA> 59584a64-3b7f-4180-86eb-cde242458f1c       1970-01-01
#> 39             <NA> 601c6c97-3141-492d-869c-024fafee1783       1970-01-01
#> 40             <NA> 5ee6ca45-445f-4219-8618-e76ed2e5a70d       1970-01-01
#> 41             <NA>                     3587011000001109       1970-01-01
#>    valid_end_date invalid_reason
#> 1      2099-12-31           <NA>
#> 2      2099-12-31           <NA>
#> 3      2099-12-31           <NA>
#> 4      2005-07-26              D
#> 5      2099-12-31           <NA>
#> 6      2018-06-03              D
#> 7      2099-12-31           <NA>
#> 8      2099-12-31           <NA>
#> 9      2099-12-31           <NA>
#> 10     2099-12-31           <NA>
#> 11     2099-12-31           <NA>
#> 12     2099-12-31           <NA>
#> 13     2017-08-23              D
#> 14     2099-12-31           <NA>
#> 15     2099-12-31           <NA>
#> 16     2099-12-31           <NA>
#> 17     2099-12-31           <NA>
#> 18     2099-12-31           <NA>
#> 19     2099-12-31           <NA>
#> 20     2099-12-31           <NA>
#> 21     2099-12-31           <NA>
#> 22     2099-12-31           <NA>
#> 23     2099-12-31           <NA>
#> 24     2099-12-31           <NA>
#> 25     2099-12-31           <NA>
#> 26     2099-12-31           <NA>
#> 27     2099-12-31           <NA>
#> 28     2099-12-31           <NA>
#> 29     2099-12-31           <NA>
#> 30     2099-12-31           <NA>
#> 31     2099-12-31           <NA>
#> 32     2099-12-31           <NA>
#> 33     2099-12-31           <NA>
#> 34     2099-12-31           <NA>
#> 35     2099-12-31           <NA>
#> 36     2099-12-31           <NA>
#> 37     2099-12-31           <NA>
#> 38     2099-12-31           <NA>
#> 39     2099-12-31           <NA>
#> 40     2099-12-31           <NA>
#> 41     2005-07-26              U
#> 
#> 

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE,
            render_sql = TRUE,
            "concept_id",
            "concept_name")

#> [2020-11-29 18:44:43]	SQL: SELECT * FROM public.concept t WHERE t.concept_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:43]	SQL: SELECT * FROM public.concept t WHERE t.concept_name::varchar IN ('Aspirin') ;

#> $ROWS
#> # A tibble: 2 x 2
#>   Field         Rows
#>   <chr>        <int>
#> 1 concept_id       0
#> 2 concept_name    41
#> 
#> $RESULTSETS
#> $RESULTSETS$concept_name
#>    concept_id concept_name   domain_id    vocabulary_id    concept_class_id
#> 1     4194196      Aspirin        Drug           SNOMED Pharma/Biol Product
#> 2     4306886      Aspirin        Drug           SNOMED           Substance
#> 3     4325480      Aspirin        Drug            NDFRT  Chemical Structure
#> 4    21197519      Aspirin        Drug             dm+d          Ingredient
#> 5    21234362      Aspirin        Drug             dm+d          Ingredient
#> 6    21272111      Aspirin        Drug             dm+d                 VTM
#> 7    35144833      Aspirin        Drug             JMDC        Drug Product
#> 8    35145382      Aspirin        Drug             JMDC        Drug Product
#> 9    35802900      Aspirin        Drug           HemOnc           Component
#> 10   36308940      Aspirin  Meas Value            LOINC              Answer
#> 11   37491813      Aspirin        Drug              CTD          Ingredient
#> 12   43368904      Aspirin        Drug              AMT        AU Substance
#> 13   43565741      Aspirin        Drug RxNorm Extension          Brand Name
#> 14   45615296      Aspirin        Drug             MeSH        Main Heading
#> 15   45627474      Aspirin        Drug              SPL                 SPL
#> 16   45633182      Aspirin        Drug              SPL                 SPL
#> 17   45639008      Aspirin        Drug              SPL                 SPL
#> 18   45639297      Aspirin        Drug              SPL                 SPL
#> 19   45639553      Aspirin        Drug              SPL                 SPL
#> 20   45643575      Aspirin        Drug              SPL                 SPL
#> 21   45645704      Aspirin        Drug              SPL                 SPL
#> 22   45649656      Aspirin        Drug              SPL                 SPL
#> 23   45652097      Aspirin        Drug              SPL                 SPL
#> 24   45661387      Aspirin        Drug              SPL                 SPL
#> 25   45662675      Aspirin        Drug              SPL                 SPL
#> 26   45662754      Aspirin        Drug              SPL                 SPL
#> 27   45663354      Aspirin        Drug              SPL                 SPL
#> 28   45665235      Aspirin        Drug              SPL                 SPL
#> 29   45671751      Aspirin        Drug              SPL                 SPL
#> 30   45672031      Aspirin        Drug              SPL                 SPL
#> 31   45673209      Aspirin        Drug              SPL                 SPL
#> 32   45678154      Aspirin        Drug              SPL                 SPL
#> 33   45685766      Aspirin        Drug              SPL                 SPL
#> 34   45687425      Aspirin        Drug              SPL                 SPL
#> 35   45689848      Aspirin        Drug              SPL                 SPL
#> 36   45690793      Aspirin        Drug              SPL                 SPL
#> 37   45692943      Aspirin        Drug              SPL                 SPL
#> 38   45693677      Aspirin        Drug              SPL                 SPL
#> 39   45693813      Aspirin        Drug              SPL                 SPL
#> 40   45699321      Aspirin        Drug              SPL                 SPL
#> 41   45957786      Aspirin Observation           SNOMED           Substance
#>    standard_concept                         concept_code valid_start_date
#> 1              <NA>                              7947003       1970-01-01
#> 2              <NA>                            387458008       2005-07-27
#> 3              <NA>                          N0000006582       1970-01-01
#> 4              <NA>                     3587011000001109       1970-01-01
#> 5              <NA>                            387458008       2005-07-27
#> 6              <NA>                              7947003       1970-01-01
#> 7              <NA>                         100000042000       1970-01-01
#> 8              <NA>                         100000071579       1970-01-01
#> 9              <NA>                                   47       2019-05-27
#> 10                S                            LA26702-3       1970-01-01
#> 11             <NA>                              50-78-2       1970-01-01
#> 12             <NA>                     1817011000036100       2016-11-01
#> 13             <NA>                           OMOP883993       1970-01-01
#> 14             <NA>                              D001241       1970-01-01
#> 15             <NA> 6c5da7fb-6e7b-4ac3-8a75-728c1e50fe1d       1970-01-01
#> 16             <NA> 7136e21e-49ad-4b7a-a9a1-7cfe3f1d80d5       1970-01-01
#> 17             <NA> 8174dd75-2e11-4d4f-aa69-2b3a33a19b0d       1970-01-01
#> 18             <NA> 90b5a964-00cb-4a1c-a19a-442bfa0f18bf       1970-01-01
#> 19             <NA> 9c0a0b9b-ef6f-4a60-acca-84dc815deae5       1970-01-01
#> 20             <NA> 5264ec86-1571-4d7a-a5d9-a52b2e4fcdf3       1970-01-01
#> 21             <NA> b528bc43-d350-4d03-a463-720e52532749       1970-01-01
#> 22             <NA> 6c584de6-2131-4d48-a3f5-18408aff9b2d       1970-01-01
#> 23             <NA> daeaa324-6546-4330-a9d3-b0b93da2d853       1970-01-01
#> 24             <NA> 87e43d62-aaa7-459f-8202-a629a4d7a423       1970-01-01
#> 25             <NA> c4efae47-dfb4-4102-a620-64631daef629       1970-01-01
#> 26             <NA> c86cffe1-949b-48b2-b5fb-388c6579de84       1970-01-01
#> 27             <NA> e4c26347-6af0-403a-81f6-4dbef95d5553       1970-01-01
#> 28             <NA> 3b8a0939-9e08-46da-8b3c-fba9c3fea9ff       1970-01-01
#> 29             <NA> 6a82cf66-a806-4331-b6a6-d13a85ec6fec       1970-01-01
#> 30             <NA> 76f12351-1855-4901-a042-b919a3539787       1970-01-01
#> 31             <NA> ac4fe306-e45d-4855-98c9-eca227b51f92       1970-01-01
#> 32             <NA> 8e161987-6226-41b7-82ef-1536ad7b1251       1970-01-01
#> 33             <NA> ee276b52-7c66-45c4-9c88-c10f47f4dbb0       1970-01-01
#> 34             <NA> 3a650603-ebcd-4239-83d3-0a4447bd288c       1970-01-01
#> 35             <NA> a9dfc258-6e3d-48fb-a6f2-b98d6f7cea47       1970-01-01
#> 36             <NA> d45d19b1-38cd-408c-a1e4-e01e7d2014ff       1970-01-01
#> 37             <NA> 37af67f9-9741-4d67-8eb6-f6f29a60f26b       1970-01-01
#> 38             <NA> 59584a64-3b7f-4180-86eb-cde242458f1c       1970-01-01
#> 39             <NA> 601c6c97-3141-492d-869c-024fafee1783       1970-01-01
#> 40             <NA> 5ee6ca45-445f-4219-8618-e76ed2e5a70d       1970-01-01
#> 41             <NA>                     3587011000001109       1970-01-01
#>    valid_end_date invalid_reason
#> 1      2099-12-31           <NA>
#> 2      2099-12-31           <NA>
#> 3      2099-12-31           <NA>
#> 4      2005-07-26              D
#> 5      2099-12-31           <NA>
#> 6      2018-06-03              D
#> 7      2099-12-31           <NA>
#> 8      2099-12-31           <NA>
#> 9      2099-12-31           <NA>
#> 10     2099-12-31           <NA>
#> 11     2099-12-31           <NA>
#> 12     2099-12-31           <NA>
#> 13     2017-08-23              D
#> 14     2099-12-31           <NA>
#> 15     2099-12-31           <NA>
#> 16     2099-12-31           <NA>
#> 17     2099-12-31           <NA>
#> 18     2099-12-31           <NA>
#> 19     2099-12-31           <NA>
#> 20     2099-12-31           <NA>
#> 21     2099-12-31           <NA>
#> 22     2099-12-31           <NA>
#> 23     2099-12-31           <NA>
#> 24     2099-12-31           <NA>
#> 25     2099-12-31           <NA>
#> 26     2099-12-31           <NA>
#> 27     2099-12-31           <NA>
#> 28     2099-12-31           <NA>
#> 29     2099-12-31           <NA>
#> 30     2099-12-31           <NA>
#> 31     2099-12-31           <NA>
#> 32     2099-12-31           <NA>
#> 33     2099-12-31           <NA>
#> 34     2099-12-31           <NA>
#> 35     2099-12-31           <NA>
#> 36     2099-12-31           <NA>
#> 37     2099-12-31           <NA>
#> 38     2099-12-31           <NA>
#> 39     2099-12-31           <NA>
#> 40     2099-12-31           <NA>
#> 41     2005-07-26              U
#> 
#> 

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            c("concept_id", "concept_name"),
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE,
            render_sql = TRUE)

#> [2020-11-29 18:44:45]	SQL: SELECT * FROM public.concept t WHERE t.concept_id::varchar IN ('Aspirin') ;
#> [2020-11-29 18:44:45]	SQL: SELECT * FROM public.concept t WHERE t.concept_name::varchar IN ('Aspirin') ;

#> $ROWS
#> # A tibble: 2 x 2
#>   Field         Rows
#>   <chr>        <int>
#> 1 concept_id       0
#> 2 concept_name    41
#> 
#> $RESULTSETS
#> $RESULTSETS$concept_name
#>    concept_id concept_name   domain_id    vocabulary_id    concept_class_id
#> 1     4194196      Aspirin        Drug           SNOMED Pharma/Biol Product
#> 2     4306886      Aspirin        Drug           SNOMED           Substance
#> 3     4325480      Aspirin        Drug            NDFRT  Chemical Structure
#> 4    21197519      Aspirin        Drug             dm+d          Ingredient
#> 5    21234362      Aspirin        Drug             dm+d          Ingredient
#> 6    21272111      Aspirin        Drug             dm+d                 VTM
#> 7    35144833      Aspirin        Drug             JMDC        Drug Product
#> 8    35145382      Aspirin        Drug             JMDC        Drug Product
#> 9    35802900      Aspirin        Drug           HemOnc           Component
#> 10   36308940      Aspirin  Meas Value            LOINC              Answer
#> 11   37491813      Aspirin        Drug              CTD          Ingredient
#> 12   43368904      Aspirin        Drug              AMT        AU Substance
#> 13   43565741      Aspirin        Drug RxNorm Extension          Brand Name
#> 14   45615296      Aspirin        Drug             MeSH        Main Heading
#> 15   45627474      Aspirin        Drug              SPL                 SPL
#> 16   45633182      Aspirin        Drug              SPL                 SPL
#> 17   45639008      Aspirin        Drug              SPL                 SPL
#> 18   45639297      Aspirin        Drug              SPL                 SPL
#> 19   45639553      Aspirin        Drug              SPL                 SPL
#> 20   45643575      Aspirin        Drug              SPL                 SPL
#> 21   45645704      Aspirin        Drug              SPL                 SPL
#> 22   45649656      Aspirin        Drug              SPL                 SPL
#> 23   45652097      Aspirin        Drug              SPL                 SPL
#> 24   45661387      Aspirin        Drug              SPL                 SPL
#> 25   45662675      Aspirin        Drug              SPL                 SPL
#> 26   45662754      Aspirin        Drug              SPL                 SPL
#> 27   45663354      Aspirin        Drug              SPL                 SPL
#> 28   45665235      Aspirin        Drug              SPL                 SPL
#> 29   45671751      Aspirin        Drug              SPL                 SPL
#> 30   45672031      Aspirin        Drug              SPL                 SPL
#> 31   45673209      Aspirin        Drug              SPL                 SPL
#> 32   45678154      Aspirin        Drug              SPL                 SPL
#> 33   45685766      Aspirin        Drug              SPL                 SPL
#> 34   45687425      Aspirin        Drug              SPL                 SPL
#> 35   45689848      Aspirin        Drug              SPL                 SPL
#> 36   45690793      Aspirin        Drug              SPL                 SPL
#> 37   45692943      Aspirin        Drug              SPL                 SPL
#> 38   45693677      Aspirin        Drug              SPL                 SPL
#> 39   45693813      Aspirin        Drug              SPL                 SPL
#> 40   45699321      Aspirin        Drug              SPL                 SPL
#> 41   45957786      Aspirin Observation           SNOMED           Substance
#>    standard_concept                         concept_code valid_start_date
#> 1              <NA>                              7947003       1970-01-01
#> 2              <NA>                            387458008       2005-07-27
#> 3              <NA>                          N0000006582       1970-01-01
#> 4              <NA>                     3587011000001109       1970-01-01
#> 5              <NA>                            387458008       2005-07-27
#> 6              <NA>                              7947003       1970-01-01
#> 7              <NA>                         100000042000       1970-01-01
#> 8              <NA>                         100000071579       1970-01-01
#> 9              <NA>                                   47       2019-05-27
#> 10                S                            LA26702-3       1970-01-01
#> 11             <NA>                              50-78-2       1970-01-01
#> 12             <NA>                     1817011000036100       2016-11-01
#> 13             <NA>                           OMOP883993       1970-01-01
#> 14             <NA>                              D001241       1970-01-01
#> 15             <NA> 6c5da7fb-6e7b-4ac3-8a75-728c1e50fe1d       1970-01-01
#> 16             <NA> 7136e21e-49ad-4b7a-a9a1-7cfe3f1d80d5       1970-01-01
#> 17             <NA> 8174dd75-2e11-4d4f-aa69-2b3a33a19b0d       1970-01-01
#> 18             <NA> 90b5a964-00cb-4a1c-a19a-442bfa0f18bf       1970-01-01
#> 19             <NA> 9c0a0b9b-ef6f-4a60-acca-84dc815deae5       1970-01-01
#> 20             <NA> 5264ec86-1571-4d7a-a5d9-a52b2e4fcdf3       1970-01-01
#> 21             <NA> b528bc43-d350-4d03-a463-720e52532749       1970-01-01
#> 22             <NA> 6c584de6-2131-4d48-a3f5-18408aff9b2d       1970-01-01
#> 23             <NA> daeaa324-6546-4330-a9d3-b0b93da2d853       1970-01-01
#> 24             <NA> 87e43d62-aaa7-459f-8202-a629a4d7a423       1970-01-01
#> 25             <NA> c4efae47-dfb4-4102-a620-64631daef629       1970-01-01
#> 26             <NA> c86cffe1-949b-48b2-b5fb-388c6579de84       1970-01-01
#> 27             <NA> e4c26347-6af0-403a-81f6-4dbef95d5553       1970-01-01
#> 28             <NA> 3b8a0939-9e08-46da-8b3c-fba9c3fea9ff       1970-01-01
#> 29             <NA> 6a82cf66-a806-4331-b6a6-d13a85ec6fec       1970-01-01
#> 30             <NA> 76f12351-1855-4901-a042-b919a3539787       1970-01-01
#> 31             <NA> ac4fe306-e45d-4855-98c9-eca227b51f92       1970-01-01
#> 32             <NA> 8e161987-6226-41b7-82ef-1536ad7b1251       1970-01-01
#> 33             <NA> ee276b52-7c66-45c4-9c88-c10f47f4dbb0       1970-01-01
#> 34             <NA> 3a650603-ebcd-4239-83d3-0a4447bd288c       1970-01-01
#> 35             <NA> a9dfc258-6e3d-48fb-a6f2-b98d6f7cea47       1970-01-01
#> 36             <NA> d45d19b1-38cd-408c-a1e4-e01e7d2014ff       1970-01-01
#> 37             <NA> 37af67f9-9741-4d67-8eb6-f6f29a60f26b       1970-01-01
#> 38             <NA> 59584a64-3b7f-4180-86eb-cde242458f1c       1970-01-01
#> 39             <NA> 601c6c97-3141-492d-869c-024fafee1783       1970-01-01
#> 40             <NA> 5ee6ca45-445f-4219-8618-e76ed2e5a70d       1970-01-01
#> 41             <NA>                     3587011000001109       1970-01-01
#>    valid_end_date invalid_reason
#> 1      2099-12-31           <NA>
#> 2      2099-12-31           <NA>
#> 3      2099-12-31           <NA>
#> 4      2005-07-26              D
#> 5      2099-12-31           <NA>
#> 6      2018-06-03              D
#> 7      2099-12-31           <NA>
#> 8      2099-12-31           <NA>
#> 9      2099-12-31           <NA>
#> 10     2099-12-31           <NA>
#> 11     2099-12-31           <NA>
#> 12     2099-12-31           <NA>
#> 13     2017-08-23              D
#> 14     2099-12-31           <NA>
#> 15     2099-12-31           <NA>
#> 16     2099-12-31           <NA>
#> 17     2099-12-31           <NA>
#> 18     2099-12-31           <NA>
#> 19     2099-12-31           <NA>
#> 20     2099-12-31           <NA>
#> 21     2099-12-31           <NA>
#> 22     2099-12-31           <NA>
#> 23     2099-12-31           <NA>
#> 24     2099-12-31           <NA>
#> 25     2099-12-31           <NA>
#> 26     2099-12-31           <NA>
#> 27     2099-12-31           <NA>
#> 28     2099-12-31           <NA>
#> 29     2099-12-31           <NA>
#> 30     2099-12-31           <NA>
#> 31     2099-12-31           <NA>
#> 32     2099-12-31           <NA>
#> 33     2099-12-31           <NA>
#> 34     2099-12-31           <NA>
#> 35     2099-12-31           <NA>
#> 36     2099-12-31           <NA>
#> 37     2099-12-31           <NA>
#> 38     2099-12-31           <NA>
#> 39     2099-12-31           <NA>
#> 40     2099-12-31           <NA>
#> 41     2005-07-26              U
#> 
#> 

dc(conn = conn,
   remove = TRUE)

#> [2020-11-29 18:44:47]	Postgres connection closed